# Add benchmark for nullable parsing