--- a/persistent-template/bench/Main.hs
+++ b/persistent-template/bench/Main.hs
@@ -47,6 +47,7 @@ main = defaultMain
                 , bench "10x100" $ nfIO $ mkPersist' $(parseReferencesQ (mkNullableModels 10 100))
                 -- , bench "10x1000" $ nfIO $ mkPersist' $(parseReferencesQ (mkNullableModels 10 1000))
                 ]
+            , bench "From File" $ nfIO $ mkPersist' $(persistFileWith lowerCaseSettings "bench/models-slowly")
             ]
         ]
     ]