… persistent-template/bench/models-slowly