Re: #824 

This set of benchmarks checks how parsing works with `Maybe` fields.

```
Benchmark persistent-th-bench: RUNNING...
benchmarking Parsing/Non-Null Fields/Increasing model count/1x10
time                 903.2 μs   (896.3 μs .. 911.7 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 911.2 μs   (906.3 μs .. 918.6 μs)
std dev              20.03 μs   (15.10 μs .. 26.60 μs)
variance introduced by outliers: 12% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing model count/10x10
time                 916.8 μs   (902.0 μs .. 929.6 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 925.9 μs   (916.3 μs .. 941.6 μs)
std dev              41.28 μs   (25.88 μs .. 68.38 μs)
variance introduced by outliers: 34% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing model count/100x10
time                 17.55 ms   (17.17 ms .. 18.00 ms)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 16.83 ms   (16.22 ms .. 17.16 ms)
std dev              1.055 ms   (571.6 μs .. 1.811 ms)
variance introduced by outliers: 26% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing model count/1000x10
time                 188.9 ms   (128.0 ms .. 244.4 ms)
                     0.971 R²   (0.907 R² .. 1.000 R²)
mean                 176.6 ms   (153.1 ms .. 193.0 ms)
std dev              23.93 ms   (10.99 ms .. 33.45 ms)
variance introduced by outliers: 37% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing field count/100x1
time                 2.628 ms   (2.608 ms .. 2.646 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 2.672 ms   (2.640 ms .. 2.728 ms)
std dev              139.0 μs   (97.65 μs .. 208.4 μs)
variance introduced by outliers: 34% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing field count/100x10
time                 16.80 ms   (16.50 ms .. 17.11 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 16.22 ms   (15.68 ms .. 16.50 ms)
std dev              905.8 μs   (469.5 μs .. 1.560 ms)
variance introduced by outliers: 22% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing field count/100x100
time                 155.0 ms   (115.3 ms .. 192.6 ms)
                     0.973 R²   (0.925 R² .. 1.000 R²)
mean                 151.5 ms   (134.2 ms .. 165.0 ms)
std dev              19.91 ms   (11.67 ms .. 29.96 ms)
variance introduced by outliers: 32% (moderately inflated)

benchmarking Parsing/Non-Null Fields/Increasing field count/100x1000
time                 1.882 s    (1.088 s .. 2.948 s)
                     0.965 R²   (0.904 R² .. 1.000 R²)
mean                 1.570 s    (1.254 s .. 1.758 s)
std dev              288.2 ms   (0.0 s .. 326.1 ms)
variance introduced by outliers: 47% (moderately inflated)

benchmarking Parsing/Nullable/Increasing model count/1x10
time                 1.024 ms   (1.020 ms .. 1.031 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.004 ms   (996.5 μs .. 1.010 ms)
std dev              21.88 μs   (17.20 μs .. 28.36 μs)
variance introduced by outliers: 11% (moderately inflated)

benchmarking Parsing/Nullable/Increasing model count/10x10
time                 1.014 ms   (1.009 ms .. 1.020 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.000 ms   (995.7 μs .. 1.004 ms)
std dev              14.17 μs   (11.28 μs .. 18.26 μs)

benchmarking Parsing/Nullable/Increasing model count/100x10
time                 19.84 ms   (19.41 ms .. 20.27 ms)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 18.74 ms   (18.05 ms .. 19.22 ms)
std dev              1.264 ms   (768.4 μs .. 1.832 ms)
variance introduced by outliers: 28% (moderately inflated)

benchmarking Parsing/Nullable/Increasing model count/1000x10
time                 215.4 ms   (146.5 ms .. 277.4 ms)
                     0.970 R²   (0.914 R² .. 1.000 R²)
mean                 197.2 ms   (167.7 ms .. 217.0 ms)
std dev              29.38 ms   (11.83 ms .. 40.81 ms)
variance introduced by outliers: 38% (moderately inflated)

benchmarking Parsing/Nullable/Increasing field count/100x1
time                 2.883 ms   (2.868 ms .. 2.899 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.832 ms   (2.815 ms .. 2.846 ms)
std dev              49.56 μs   (41.42 μs .. 60.00 μs)

benchmarking Parsing/Nullable/Increasing field count/100x10
time                 19.89 ms   (19.48 ms .. 20.30 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 18.80 ms   (18.07 ms .. 19.14 ms)
std dev              1.113 ms   (580.2 μs .. 1.888 ms)
variance introduced by outliers: 24% (moderately inflated)

benchmarking Parsing/Nullable/Increasing field count/100x100
time                 193.2 ms   (130.4 ms .. 249.8 ms)
                     0.970 R²   (0.913 R² .. 1.000 R²)
mean                 177.7 ms   (151.7 ms .. 195.6 ms)
std dev              26.08 ms   (12.24 ms .. 36.23 ms)
variance introduced by outliers: 38% (moderately inflated)

benchmarking Parsing/Nullable/Increasing field count/100x1000
time                 2.195 s    (1.341 s .. 3.400 s)
                     0.968 R²   (0.914 R² .. 1.000 R²)
mean                 1.821 s    (1.451 s .. 2.037 s)
std dev              333.9 ms   (0.0 s .. 375.0 ms)
variance introduced by outliers: 47% (moderately inflated)
```

These benchmarks demonstrate that adding `Maybe` to the end does result in a small but significant increase of time to parse.

---

Before submitting your PR, check that you've:

- [ ] Bumped the version number
- [ ] Documented new APIs with [Haddock markup](https://www.haskell.org/haddock/doc/html/index.html)
- [ ] Added [`@since` declarations](http://haskell-haddock.readthedocs.io/en/latest/markup.html#since) to the Haddock

After submitting your PR:

- [ ] Update the Changelog.md file with a link to your PR
- [ ] Check that CI passes (or if it fails, for reasons unrelated to your change, like CI timeouts)

<!---Thanks so much for contributing! :)

_If these checkboxes don't apply to your PR, you can delete them_-->